backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic[email]==2.5.2
python-dotenv==1.0.0
requests==2.31.0
